Character.javaをCheckSentence.javaとFirstChar.javaに分けました
