+++ src/project/drawwordchain/model/CheckSentence.java
@@ -1,16 +1,6 @@
 package project.drawwordchain.model;
 
-import java.util.Random;
-
-public class Character {
-
-  //50音の[を]と[ん]を抜いた文字。計４４文字
-  String[] charList = {"あ","い","う","え","お","か","き",
-    "く","け","こ","さ","し","す","せ","そ","た","ち","つ",
-    "て","と","な","に","ぬ","ね","の","は","ひ","ふ","へ",
-    "ほ","ま","み","む","め","も","や","ゆ","よ","ら","り",
-    "る","れ","ろ","わ"};
-
+public class CheckSentence {
   //濁音、半濁音、促音、拗音のリスト。計３０文字
   char[] sonantList = {'が','ぎ','ぐ','げ','ご','ざ','じ',
     'ず','ぜ','ぞ','だ','ぢ','づ','で','ど','ば','び','ぶ','べ',
@@ -20,17 +10,6 @@ public class Character {
     'さ','し','す','せ','そ','た','ち','つ','て','と','は',
     'ひ','ふ','へ','ほ','は','ひ','ふ','へ','ほ','や','ゆ',
     'よ','わ','つ'};
-
-//しりとり開始時の最初の文字を決定するメソッド
-  public String getChar(){
-    //Randomクラスの生成
-    Random r = new Random();
-    //乱数取得
-    int i = r.nextInt(44);
-    //charListの文字をランダムにreturnする
-    return charList[i];
-  }
-}
 
   //しりとりの一つ目の単語と二つ目の単語がしりとりとしての正否を判定するメソッド
   //引数(一つ目の単語,二つ目の単語)
@@ -60,3 +39,5 @@ public class Character {
 	  else return false;
 
   }
+
+}
